update: reuse types that can be shared
