+++ protocol-units/bridge/chains/ethereum/src/event_monitoring.rs
@@ -1,5 +1,4 @@
 use crate::counterparty_contract::SmartContractCounterpartyEvent;
-use crate::initiator_contract::SmartContractInitiatorEvent;
 use crate::types::{
 	EthAddress, EthHash, EventName, COUNTERPARTY_ABORTED_SELECT, COUNTERPARTY_COMPLETED_SELECT,
 	COUNTERPARTY_LOCKED_SELECT, INITIATOR_COMPLETED_SELECT, INITIATOR_INITIATED_SELECT,
@@ -18,7 +17,7 @@ use alloy::{
 use bridge_shared::bridge_monitoring::{
 	BridgeContractCounterpartyEvent, BridgeContractCounterpartyMonitoring,
 };
-use bridge_shared::types::{Amount, LockDetails, TimeLock};
+use bridge_shared::types::{Amount, LockDetails, SmartContractInitiatorEvent, TimeLock};
 use bridge_shared::{
 	bridge_monitoring::{BridgeContractInitiatorEvent, BridgeContractInitiatorMonitoring},
 	types::{
